requirements.txt
Flask==3.0.0
pypdf==3.17.0
google-generativeai==0.3.2
python-dotenv==1.0.0